backend/requirements.txt
fastapi==0.110.0
langchain_community==0.0.27
langchain_openai==0.0.8
langchain-core==0.3.12 
langchain_aws==0.2.2
langchain_mongodb==0.2.0
langchain==0.1.9
openai==1.13.3
Pillow==10.2.0
pymongo==4.6.2
torchvision==0.17.1
uvicorn==0.27.1
python-dotenv==1.0.1